requirements.txt
flask
flask-sqlalchemy
flask-session
argon2-cffi
python-dotenv
